refactor(prestashop): move RetryConfig/PrestashopWebserviceClientOptions to the interface file
Follow-up on the tech-review SUGGESTION about type placement: these
were declared inline in prestashop-webservice.client.ts, deviating
from engineering-standards.md's "types in separate *.types.ts files"
rule. Moved both to prestashop-webservice.client.interface.ts
alongside the client's other supporting types (PrestashopQueryFilters,
PrestashopWriteOptions) - the file already plays that role for this
client, so no new file is needed.
