+++ libs/integrations/prestashop/src/infrastructure/http/prestashop-webservice.client.interface.ts
@@ -7,6 +7,32 @@
  *
  * @module libs/integrations/prestashop/src/infrastructure/http
  */
+import type { FetchLike } from '@openlinker/shared/http';
+
+/**
+ * Retry configuration for the WebService client's transport.
+ */
+export interface RetryConfig {
+  maxRetries: number;
+  initialDelayMs: number;
+  maxDelayMs: number;
+  backoffMultiplier: number;
+}
+
+/**
+ * Optional constructor knobs, grouped so adding a future option never forces
+ * every caller to thread an `undefined` positional placeholder through the
+ * ones it doesn't need (#1810 review follow-up).
+ */
+export interface PrestashopWebserviceClientOptions {
+  retryConfig?: Partial<RetryConfig>;
+  /**
+   * Connection-bound outbound transport (#1810) — defaults to bare fetch so
+   * existing call sites that predate the rate-limit mechanism (test
+   * harnesses, other-package callers) keep working unchanged.
+   */
+  fetchImpl?: FetchLike;
+}
 
 /**
  * PrestaShop query filters for list operations
